refactor: change places from usual structure to feature-slice
