+++ src/app/App.tsx
@@ -1,15 +1,15 @@
 import { Suspense } from "react"
 import "./styles/index.scss"
 import { Routes, Route, Link } from "react-router-dom"
-import { MainPageAsync } from "./pages/MainPage/MainPage.async"
-import { AboutPageAsync } from "./pages/AboutPage/AboutPage.async"
-import { useTheme } from "./theme/useTheme"
-import { classNames } from "./helpers/classNames/classNames"
+import { classNames } from "../helpers/classNames/classNames"
+import { useTheme } from "./provider/ThemeProvider/lib/useTheme"
+import { MainPageAsync } from "../pages/MainPage/MainPage.async"
+import { AboutPageAsync } from "../pages/AboutPage/AboutPage.async"
 
 const App = () => {
-  const { theme, toggleTheme } = useTheme();
+  const { theme, toggleTheme } = useTheme()
   return (
-    <div className={classNames('app', {}, [theme])}>
+    <div className={classNames("app", {}, [theme])}>
       <button onClick={toggleTheme}>Кнопелла</button>
       <Link to={"/"}>Главная</Link>
       <Link to={"/about"}>О сайте</Link>
