+++ src/app/provider/ThemeProvider/lib/useTheme.ts
@@ -1,4 +1,8 @@
-import { LOCAL_STORAGE_THEME_KEY, Theme, ThemeContext } from "./Theme.context"
+import {
+  LOCAL_STORAGE_THEME_KEY,
+  Theme,
+  ThemeContext,
+} from "../../../../shared/config/theme/Theme.context"
 import { useContext } from "react"
 
 interface UseThemeResult {
