+++ src/app/provider/ThemeProvider/ui/ThemeProvider.tsx
@@ -1,21 +1,24 @@
 import React, { FC, useMemo, useState } from "react"
-import { LOCAL_STORAGE_THEME_KEY, Theme, ThemeContext } from "./Theme.context"
+import {
+  LOCAL_STORAGE_THEME_KEY,
+  Theme,
+  ThemeContext,
+} from "../../../../shared/config/theme/Theme.context"
 
 const defaultTheme = (localStorage.getItem(LOCAL_STORAGE_THEME_KEY) as Theme) || Theme.LIGHT
 
 const ThemeProvider: FC = ({ children }) => {
   const [theme, setTheme] = useState<Theme>(defaultTheme)
 
-  const defaultProps = useMemo(() => ({
+  const defaultProps = useMemo(
+    () => ({
       theme: theme,
       setTheme: setTheme,
-  }), [theme])
-
-  return (
-    <ThemeContext.Provider value={defaultProps}>
-      {children}
-    </ThemeContext.Provider>
+    }),
+    [theme],
   )
+
+  return <ThemeContext.Provider value={defaultProps}>{children}</ThemeContext.Provider>
 }
 
 export default ThemeProvider
