--- src/components/Counter.module.scss
@@ -1,4 +1,0 @@
-.button {
-  padding: 20px;
-  color: green;
-}
