+++ src/index.tsx
@@ -1,7 +1,7 @@
 import { render } from "react-dom"
-import App from "./App"
 import { BrowserRouter } from "react-router-dom"
-import ThemeProvider from "./theme/ThemeProvider"
+import ThemeProvider from "./app/provider/ThemeProvider/ui/ThemeProvider"
+import App from "./app/App"
 
 render(
   <BrowserRouter>
